+++ packages/graphql-kimera/src/constants.js
@@ -11,6 +11,10 @@ const interfaceTypeDefinition = 'InterfaceTypeDefinition';
 // How many items will be mocked in List Types
 const DEFAULT_LIST_LENGTH = 3;
 
+// Used to track a potentially recursive branch so we can warn the user.
+const RECURSIVITY_DEPTH_LIMIT =
+  Number(process.env.KIMERA_RECURSIVITY_DEPTH_LIMIT) || 20;
+
 module.exports = {
   ID,
   string,
@@ -20,4 +24,5 @@ module.exports = {
   scalarTypeDefinition,
   interfaceTypeDefinition,
   DEFAULT_LIST_LENGTH,
+  RECURSIVITY_DEPTH_LIMIT,
 };